Added helpful cog and moved command to it
This established the original command categories from before the rewrite
