+++ cogs/misc.py
@@ -29,24 +29,6 @@ class Misc(commands.Cog, Utils):
             await ctx.message.delete()
             await ctx.send(embed=embed)
 
-    @commands.command(name = "gamexplain",
-                      description = "Gives an explanation to how the Emu Bot game works",
-                      aliases = ['gx', 'gameexplain', 'gxplain'],
-                      brief = "Gives an explanation to how the Emu Bot game works",
-                      help = "Gives an in-depth explain of how to use the Emu Bot's game commands and how the Emu Bot game works"
-)
-    async def gamexplain(self, ctx):
-        embed = discord.Embed(title='Game Explanation', description='''**In the emu bot game, you gain credits by participating in chat. You can spend these credits on emus and attack your friends with them.**
--You earn credits by sending messages.
--You can use those credits to buy emus
--The maximum amount of emus you can have is {} emus
--You can put emus on defense
--The maximum amount of emus on defense is {} emus
--You can also use emus to attack your friends
--The maximum amount of emus you can attack with at a time is {} emus
--The amount of emus you attack someone with that go over the amount of emus they have on defense grants you 700 credits for each emu.'''.format(self.MAXEMUS, self.MAXDEFENSE, self.MAXATTACK), color=0x00ff00)
-        await ctx.send(embed = embed)
-        
     @commands.command(name = "helpersay",
                       aliases = ["hsay", "hs"],
                       hidden = True
